[FEAT] 관리자 플레이 리스트 기능 추가

+++ src/api/playlist/index.ts
@@ -14,14 +14,20 @@ export async function fetchPlaylists(lastPlaylistId: number) {
   return data;
 }
 
-// 플레이리스트 추가를 위한 POST API
-export async function fetchCreatePlaylist(title: string, movieIds: number[]) {
-  const { data } = await apiClient.post("/admin/playlist", {
-    title,
-    movieIds,
-  });
-
-  return data;
+export async function fetchCreatePlaylist(movieIds: number[], title: string) {
+  try {
+    console.log("API 요청 데이터:", { title, movieIds });
+    console.log("movieIds 배열:", movieIds);
+    const response = await apiClient.post("/admin/playlist", {
+      movieIds, // JSON 형식에 맞게 전달
+      title,
+    });
+    console.log("API 응답 데이터:", response.data);
+    return response.data;
+  } catch (error) {
+    console.error("플레이리스트 생성 요청 중 오류 발생:", error);
+    throw error; // 상위에서 에러 처리
+  }
 }
 
 // 플레이리스트 업데이트를 위한 PATCH API
@@ -41,9 +47,7 @@ export async function fetchUpdatePlaylist(
 
 // 영화 삭제
 export async function fetchDeletePlaylist(playlistId: number) {
-  const { data } = await apiClient.delete(
-    `/admin/playlist/${playlistId}`
-  );
+  const { data } = await apiClient.delete(`/admin/playlist/${playlistId}`);
   return data;
 }
 
@@ -53,16 +57,12 @@ export async function fetchCallPlaylist(
   createdAt?: string,
   size: number = 10
 ) {
-  const params = new URLSearchParams();
-
-  if (lastMovieId) {
-    params.append("last-movie-id", lastMovieId.toString());
-  }
-  if (createdAt) {
-    params.append("created-at", createdAt);
-  }
-  params.append("size", size.toString());
-
-  const { data } = await apiClient.get(`/admin/playlist/movies?${params.toString()}`);
+  const { data } = await apiClient.get("/admin/playlist/movies", {
+    params: {
+      ...(lastMovieId && { "last-movie-id": lastMovieId }),
+      ...(createdAt && { "created-at": createdAt }),
+      size,
+    },
+  });
   return data;
 }
